# backend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Graphfolio Backend API
  description: Backend API for Graphfolio - Company network visualization
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
...
  /api/search:
    get:
      tags:
      - search
      summary: Search
      description: 'Unified search endpoint.

        Returns results for stocks, podcasts, episodes, and tags matching the query.'
      operationId: search_api_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: Search query
          title: Q
        description: Search query
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          description: Max results per category
          default: 5
          title: Limit
        description: Max results per category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/search/suggest:
    get:
      tags:
      - search
      summary: Suggest
      description: 'Fast typeahead suggestions.

        Returns instant suggestions for autocomplete. Target <50ms.'
      operationId: suggest_api_search_suggest_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: Prefix query
          title: Q
        description: Prefix query
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 1
          default: 8
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/recommendations/by-ticker/{ticker}:
    get:
      tags:
      - recommendations
      summary: Get Recommendations By Ticker
      description: 'Get recommendations for the given ticker in the date range.

        Default timeframe: last 7 days.


        CDN Cache: 5 minutes'
      operationId: get_recommendations_by_ticker_api_recommendations_by_ticker__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'ISO date (YYYY-MM-DD); default: today − 7 days'
          title: Start Date
        description: 'ISO date (YYYY-MM-DD); default: today − 7 days'
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'ISO date (YYYY-MM-DD); default: today'
          title: End Date
        description: 'ISO date (YYYY-MM-DD); default: today'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Get Recommendations By Ticker Api Recommendations
                  By Ticker  Ticker  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/recommendations/by-podcaster/{podcaster_name}:
    get:
      tags:
      - recommendations
      summary: Get Recommendations By Podcaster
      description: 'Get recommendations from the given podcaster in the date range.

        Default timeframe: last 7 days. Pass podcast_slug to also match by episode_id
        (e.g. slug in episode_id).


        CDN Cache: 5 minutes'
      operationId: get_recommendations_by_podcaster_api_recommendations_by_podcaster__podcaster_name__get
      parameters:
      - name: podcaster_name
        in: path
        required: true
        schema:
          type: string
          title: Podcaster Name
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'ISO date (YYYY-MM-DD); default: today − 7 days'
          title: Start Date
        description: 'ISO date (YYYY-MM-DD); default: today − 7 days'
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'ISO date (YYYY-MM-DD); default: today'
          title: End Date
        description: 'ISO date (YYYY-MM-DD); default: today'
      - name: podcast_slug
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional slug; also matches episode_id ILIKE %slug%
          title: Podcast Slug
        description: Optional slug; also matches episode_id ILIKE %slug%
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Get Recommendations By Podcaster Api Recommendations
                  By Podcaster  Podcaster Name  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/recommendations/buzz:
    get:
      tags:
      - recommendations
      summary: Get Buzz
      description: 'Get most-discussed tickers in the last `days` days.

        Returns ticker, count, sentiment_score (avg), last_mentioned.


        CDN Cache: 5 minutes'
      operationId: get_buzz_api_recommendations_buzz_get
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Number of days to look back
          default: 30
          title: Days
        description: Number of days to look back
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Max number of tickers to return
          default: 10
          title: Limit
        description: Max number of tickers to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Get Buzz Api Recommendations Buzz Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SearchResponse:
      properties:
        stocks:
          items:
            $ref: '#/components/schemas/SearchResultItem'
          type: array
          title: Stocks
          default: []
        podcasts:
          items:
            $ref: '#/components/schemas/SearchResultItem'
          type: array
          title: Podcasts
          default: []
        episodes:
          items:
            $ref: '#/components/schemas/SearchResultItem'
          type: array
          title: Episodes
          default: []
        tags:
          items:
            $ref: '#/components/schemas/SearchResultItem'
          type: array
          title: Tags
          default: []
      type: object
      title: SearchResponse
    SearchResultItem:
      properties:
        id:
          type: string
          title: Id
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        link:
          type: string
          title: Link
        metadata:
          anyOf:
          - type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - id
      - type
      - title
      - link
      title: SearchResultItem
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer